backend/requirements.txt
Django==5.2.1
python-dotenv==1.1.0
psycopg2-binary==2.9.10
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
pillow==11.2.1
drf-extra-fields==3.7.0
